Log visible maddrs in both run_{trainer,training_monitor}.py with color
