+++ examples/albert/utils.py
@@ -1,9 +1,11 @@
 from typing import Dict, List, Tuple
 
+from pydantic import BaseModel, StrictFloat, confloat, conint
+
+import hivemind
 from hivemind.dht.crypto import RSASignatureValidator
 from hivemind.dht.schema import BytesWithPublicKey, SchemaValidator
 from hivemind.dht.validation import RecordValidatorBase
-from pydantic import BaseModel, StrictFloat, confloat, conint
 
 
 class LocalMetrics(BaseModel):
@@ -23,3 +25,14 @@ def make_validators(experiment_prefix: str) -> Tuple[List[RecordValidatorBase], 
     validators = [SchemaValidator(MetricSchema, prefix=experiment_prefix),
                   signature_validator]
     return validators, signature_validator.local_public_key
+
+
+class TextStyle:
+    BOLD = '\033[1m'
+    BLUE = '\033[34m'
+    RESET = '\033[0m'
+
+
+def format_visible_maddrs(dht: hivemind.DHT) -> None:
+    initial_peers_str = ' '.join(str(addr) for addr in dht.get_visible_maddrs())
+    return f"{TextStyle.BOLD}{TextStyle.BLUE}--initial_peers {initial_peers_str}{TextStyle.RESET}"
